dataset-explorer/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
httpx
python-jose
passlib[bcrypt]
sentence-transformers
scikit-learn
numpy 